+++ src/app/main/letter-box/layout.tsx
@@ -1,21 +1,19 @@
-'use client';
-
+import type {Metadata} from 'next';
 import type {ReactNode} from 'react';
 
-import {LayoutBackGround} from '@components/atom';
+import {Layout} from '@feature/letterBox/components/template';
 
 type LetterBoxLayoutProps = {
   children: ReactNode;
 };
 
+export const metadata: Metadata = {
+  title: '바다편지(우편함)',
+};
+
 const LetterBoxLayout = (props: LetterBoxLayoutProps) => {
   const {children} = props;
-  return (
-    <>
-      <title>마음을 담다(우편함)</title>
-      <LayoutBackGround>{children}</LayoutBackGround>
-    </>
-  );
+  return <Layout>{children}</Layout>;
 };
 
 export default LetterBoxLayout;
