+++ src/app/main/trash-box/layout.tsx
@@ -1,21 +1,19 @@
-'use client';
-
+import type {Metadata} from 'next';
 import type {ReactNode} from 'react';
 
-import {LayoutBackGround} from '@components/atom';
+import {Layout} from '@feature/trash/components/template';
 
 type ThrashLayoutProps = {
   children: ReactNode;
 };
 
+export const metadata: Metadata = {
+  title: '바다편지(휴지통)',
+};
+
 const TrashLayout = (props: ThrashLayoutProps) => {
   const {children} = props;
-  return (
-    <>
-      <title>마음을 담다(휴지통)</title>
-      <LayoutBackGround>{children}</LayoutBackGround>
-    </>
-  );
+  return <Layout>{children}</Layout>;
 };
 
 export default TrashLayout;
