+++ src/ui/feature/letter/read/components/template/index.ts
@@ -1,1 +1,2 @@
 export {default as Read} from './page';
+export {default as Layout} from './page/Layout';
