+++ src/ui/feature/letter/writing/components/template/index.ts
@@ -1,1 +1,2 @@
 export {default as Writing} from './page';
+export {default as Layout} from './page/Layout';
