+++ src/ui/feature/letterBox/components/template/index.ts
@@ -1,2 +1,3 @@
 export {default as LetterBox} from './page';
+export {default as Layout} from './page/Layout';
 export {default as MyLetterDetail} from './page/MyLetter';
