+++ src/ui/feature/main/components/template/index.tsx
@@ -1,1 +1,2 @@
 export {default as Main} from './page';
+export {default as Layout} from './page/Layout';
