+++ src/ui/feature/notice/components/template/index.ts
@@ -1,1 +1,2 @@
 export {default as Notice} from './page';
+export {default as Layout} from './page/Layout';
