+++ src/ui/feature/reply/components/template/index.ts
@@ -1,2 +1,3 @@
 export {default as ReplyBox} from './replyBox';
 export {default as Reply} from './reply';
+export {default as Layout} from './Layout';
