+++ src/ui/feature/setting/components/template/index.ts
@@ -1,2 +1,3 @@
 export {default as Setting} from './page/setting';
 export {default as Email} from './page/email';
+export {default as Layout} from './page/Layout';
