+++ src/ui/feature/trash/components/template/Layout.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import type {ReactNode} from 'react';
+
+import {LayoutBackGround} from '@components/atom';
+
+type ThrashLayoutProps = {
+  children: ReactNode;
+};
+
+const TrashLayout = (props: ThrashLayoutProps) => {
+  const {children} = props;
+  return <LayoutBackGround>{children}</LayoutBackGround>;
+};
+
+export default TrashLayout;
