+++ src/ui/feature/trash/components/template/index.ts
@@ -1,2 +1,3 @@
 export {default as Trash} from './trash';
 export {default as TrashBox} from './trashBox';
+export {default as Layout} from './Layout';
